+++ exthost/action_network_bandwidth.go
@@ -58,7 +58,7 @@ func getNetworkLimitBandwidthDescription() action_kit_api.ActionDescription {
 				Type:        action_kit_api.ActionParameterTypeStringArray,
 				Required:    new(false),
 				Advanced:    new(true),
-				Order:       new(105),
+				Order:       new(106),
 			},
 		),
 	}